Write a value to a namespace only if the value has changed.

Write a value to a namespace only if the value has changed.

flow_ns_sink(x, var_name, ns, flow_options = get_flow_options())

Arguments

x
Value to assign.

var_name
The name (as string) of the variable.

ns
The name space, either an environment or a Shiny::reactiveValues object.

flow_options
List of options created using get_flow_options. All options except excluded_arg and eddy are ignored.

Value

The flow object

Examples

fn <- function(x, y) { x + y }
flowed_fn <- flow_fn(1, 2, fn = fn)

#> New cache: fn=fn / fn_id=1 / fn_key=20ab2bd9a4e01c86

sunk_flow <- flow_ns_sink(flowed_fn, "test_sink_flow", new.env())