flow options used to initialize or update an eddy.

flow options used to initialize or update an eddy.

default flow options used to initialize or update an eddy.

default_flow_options(excluded_arg = character(), split_bare_list = TRUE,
  split_dataframe = FALSE, split_fn = NULL)

set_flow_options(excluded_arg = NULL, split_bare_list = NULL,
  split_dataframe = NULL, split_fn = NULL, eddy = get_current_eddy())

get_flow_options(excluded_arg = NULL, eval_arg_fn = NULL,
  split_bare_list = NULL, split_dataframe = NULL, split_fn = NULL,
  eddy = get_current_eddy())

Arguments

excluded_arg
A vector of argument names to be excluded when computing the input hash. Best used to exclude certain arguments that depend on the running state, e.g. a Shiny session, a parallel cluster, etc. Excluded arguments must have a default value to permit lazy computations. The default is not to exclude any arguments from the input hash.

split_bare_list
If the function output is a bare list (rlang::is_bare_list), determines whether to calculate the hash of each list element and create corresponding flow elements.

split_dataframe
If the function output is a data.frame or tibble, determines whether to calculate the hash of each column and create corresponding flow elements.

split_fn
Custom function to generate a list of elements from the output of the flow-ed function. Useful only if the output is not a list but a flow elements are still desired. Consider returning a list as output before using this option. If an split_fn is provided, split_bare_list and split_dataframe will be ignored.

eddy
Eddy to apply / retrieve options to / from.

eval_arg_fn
Custom function to parse the input arguments and create a list of evaluated arguments to be hashed. This function should have the exact same arguments as the original function. Try to use excluded_arg or flow source before creating a custom function. Because each custom function is flow specific, it is not possible to set this option at the eddy level using set_flow_options.

Value

For default_flow_options, a list of options.

For set_flow_options, NULL.

For get_flow_options, a list of options including the eddy.

Details

If used in set_flow_options, these options will be stored in eddy and retrieved by each flow subsequently executed. E.g. if not careful, it is possible to force all following flows to use a custom split_fn function, which is not recommended.

Sets a list of default options.

set_flow_options does not overwrite the current options when the argument is NULL.